chore: Drop dependency on fact_helper_file
The only benefit it provides over the python package magic is that
it includes some additional magic.
Now that this file is in its own repo we can simple use the
MAGIC environment variable.
As the fact_extractor is only ever run in docker it suffices to set
this in the dockerfile.